Create DavidMellis
At first glance, BlinkWithoutDelay  does not seem to be a very interesting sketch.  It looks like just a more complicated way to blink a LED.  However, BinkWithoutDelay illustrates a very important concept known as a State Machine.  

Instead of relying on delay() to time the blinking.  BlinkWithoutDelay remembers the current state of the LED and the last time it changed.  On each pass through the loop, it looks at the millis() clock to see if it is time to change the state of the LED again.

https://learn.adafruit.com/multi-tasking-the-arduino-part-1/using-millis-for-timing